Edited Player Script and level 1
Added the attacking sequence. Left click to attack
To do this, I added a hitbox child in the player prefab. The player control script will look through the children of the prefab and find "sword" This will then be the hitbox

Level 1 enemies will now follow the player and can be hit and killed