Restructured Code for Clarity and Put Level Advancement in Place
Globals now has a public bool and ushort attached to it. You'll need  to mark the bool as true for levels and set the ushort to the level number to advance properly (ex. Level 2 is 2). I'll make a seperate commit where I set the globals to minimize any sort of merge conflicts.