Read the Desription
Prefabs changed:
-Golem
-FireGoblin

Scripts changed:
-Menu
-Globals
-Portals
Why I made changes to the above scripts:
I made it so that current level and max level will increase as soon as the player touches the portal because the "Next Level" Button doesnt work for me. So when I go to Level Select, it opens level the next level for the player.